bot/requirements.txt
# aiogram 3.26 требует pydantic<2.13. Для Python 3.14 на Windows явно ставим pydantic-core
# с готовым wheel (иначе pip тянет старый core и пытается собрать из исходников — нужен Rust).
aiogram==3.26.0
pydantic==2.12.5
pydantic-core==2.41.5
python-dotenv>=1.0.0
aiohttp-socks>=0.9.0,<0.11
